config-service/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
requests-unixsocket==0.3.0